requirements.txt
opencv-python-headless
numpy
pillow
